\name{bpi_challenge_2019_test1}
\alias{bpi_challenge_2019_test1}
\title{BPI 2019 challenge test}
\usage{
bpi_challenge_2019_test1()
}
\value{
None
}
\description{
A single sales process flow from the BPI 2019 challenge event log was taken to perform the Sequence Detector testing.
The results are available in [1].
}
\references{
[1] D. Krleža, B. Vrdoljak, and M. Brčić, Latent Process Discovery using Evolving Tokenized Transducer, \emph{IEEE Access}, vol. 7, pp. 169657 - 169676, Dec. 2019
}
