\name{c_to_string}
\alias{c_to_string}
\title{String list formatting function}
\usage{
c_to_string(var)
}
\arguments{
  \item{var}{(list) - A string list}
}
\value{
(character) - An output string made of the input list elements, formatted as \emph{[e1,e2,...,en]}.
}
\description{
A method that formats an input list made of strings into a single output string.
The output string is formatted as \emph{[e1,e2,...,en]}.
}
